[CAN-54] api 호출 로직 분리
